services/id-ocr-service/requirements.txt

python-multipart==0.0.9
fastapi==0.95.2
uvicorn[standard]==0.23.2
pillow==10.0.0
pytesseract==0.3.10
openai>=1.0.0


